git-svn-id: https://mindboards.svn.sourceforge.net/svnroot/mindboards/lms_nbcnxc_128/trunk@6 c9361245-7fe8-9947-84e8-057757c4e366

+++ AT91SAM7S256/Source/c_ui.h
@@ -36,7 +36,7 @@
 
 #define   CHECKBYTE                     0x78        // Used to validate NVData
 
-#define   BATTERY_COUNT_TO_MV           13.848      // Battery count to mV factor             [mV/cnt]
+#define   BATTERY_COUNT_TO_MV    (float)13.848      // Battery count to mV factor             [mV/cnt]
 #define   LOW_BATT_THRESHOLD            6           // Low batt conunts before warning
 
 #define   BUTTON_DELAY_TIME             800         // Delay before first repeat              [mS]
